+++ integrations/gitlab/gitlab_integration/core/async_fetcher.py
@@ -6,7 +6,14 @@ from typing import List, Union, Callable, AsyncIterator, TypeVar, Any, Dict
 import gitlab.exceptions
 from gitlab import GitlabList
 from gitlab.base import RESTObject, RESTObjectList
-from gitlab.v4.objects import Project, ProjectPipelineJob, ProjectPipeline, Issue, Group
+from gitlab.v4.objects import (
+    Project,
+    ProjectPipelineJob,
+    ProjectPipeline,
+    Issue,
+    Group,
+    ProjectFile,
+)
 from loguru import logger
 
 from port_ocean.core.models import Entity
@@ -39,6 +46,7 @@ class AsyncFetcher:
         Issue,
         Project,
         Group,
+        ProjectFile,
     ]:
         with ThreadPoolExecutor() as executor:
             return await get_event_loop().run_in_executor(executor, fetch_func, *args)