+++ integrations/gitlab/CHANGELOG.md
@@ -12,7 +12,7 @@ this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### Improvements
 
-- Added support for exporting files (0.2.0)
+- Added support for exporting files (0.1.100)
 
 
 0.1.99 (2024-08-05)